Extract DashboardItem
